# backend/docs/swagger.yaml
basePath: /
definitions:
...
  db.Measurement:
    properties:
      age:
        type: number
      childrenID:
        type: string
      headCircumference:
        type: number
      height:
        type: number
      id:
        type: string
      measuredAt:
        type: string
      measurerID:
        type: string
      measurerRole:
        $ref: '#/definitions/db.UserRole'
      stuntingStatus:
        $ref: '#/definitions/db.NullStuntingStatus'
      upperArmCircumference:
        type: number
      weight:
        type: number
      zscore:
        type: number
    type: object
...
  db.NullStuntingStatus:
    properties:
      stuntingStatus:
        $ref: '#/definitions/db.StuntingStatus'
      valid:
        description: Valid is true if StuntingStatus is not NULL
        type: boolean
    type: object
...
  db.StuntingStatus:
    enum:
    - severely_stunted
    - stunted
    - normal
    - tall
    type: string
    x-enum-varnames:
    - StuntingStatusSeverelyStunted
    - StuntingStatusStunted
    - StuntingStatusNormal
    - StuntingStatusTall
  db.UserRole:
    enum:
    - tenaga_kesehatan
    - kader
    - orang_tua
    - admin
    type: string
    x-enum-varnames:
    - UserRoleTenagaKesehatan
    - UserRoleKader
    - UserRoleOrangTua
    - UserRoleAdmin
...
  handler.CreateMeasurementRequest:
    properties:
      children_id:
        type: string
      head_circumference:
        type: number
      height:
        type: number
      upper_arm_circumference:
        type: number
      weight:
        type: number
    required:
    - children_id
    - height
    - weight
    type: object
...
  handler.ErrorResponse:
    properties:
      error:
        example: error message
        type: string
    type: object
  handler.MessageResponse:
    properties:
      message:
        example: operation successful
        type: string
    type: object
...
  handler.UpdateMeasurementRequest:
    properties:
      head_circumference:
        type: number
      height:
        type: number
      upper_arm_circumference:
        type: number
      weight:
        type: number
    required:
    - height
    - weight
    type: object
...
info:
  contact: {}
  description: Backend REST API for Centing (Stunting Prevention & Monitoring System).
  title: Centing API
  version: "1.0"
paths:
...
  /nakes/children/{id}/measurements:
    get:
      description: Retrieve all growth measurements recorded for a specific child
      parameters:
      - description: Child ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/db.Measurement'
            type: array
        "400":
          description: Invalid child ID
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List measurements for a child
      tags:
      - Measurements
...
  /nakes/measurements:
    get:
      description: Retrieve all measurements recorded by authenticated healthcare
        worker
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/db.Measurement'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: 'Forbidden: insufficient permissions'
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get measurements by measurer (Nakes only)
      tags:
      - Measurements
    post:
      consumes:
      - application/json
      description: Record a new child growth measurement including weight, height,
        head and arm circumference
      parameters:
      - description: Create Measurement Payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.CreateMeasurementRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/db.Measurement'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: 'Forbidden: insufficient permissions'
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create measurement (Nakes only)
      tags:
      - Measurements
  /nakes/measurements/{id}:
    delete:
      description: Delete a growth measurement record by ID
      parameters:
      - description: Measurement ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.MessageResponse'
        "400":
          description: Invalid measurement ID
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: 'Forbidden: insufficient permissions'
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete measurement (Nakes only)
      tags:
      - Measurements
    get:
      description: Retrieve single measurement record by ID
      parameters:
      - description: Measurement ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/db.Measurement'
        "400":
          description: Invalid measurement ID
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: 'Forbidden: insufficient permissions'
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get measurement by ID (Nakes only)
      tags:
      - Measurements
    put:
      consumes:
      - application/json
      description: Update an existing growth measurement record
      parameters:
      - description: Measurement ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Update Measurement Payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateMeasurementRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/db.Measurement'
        "400":
          description: Validation error or invalid ID
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: 'Forbidden: insufficient permissions'
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update measurement (Nakes only)
      tags:
      - Measurements
...
  /ortu/children/{id}/measurements:
    get:
      description: Retrieve all growth measurements recorded for a specific child
      parameters:
      - description: Child ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/db.Measurement'
            type: array
        "400":
          description: Invalid child ID
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List measurements for a child
      tags:
      - Measurements
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"